Init StartScreen backend support on test server.
Notable changes:
-The ClientState objects now store a reference to each
connection's websocket object. This allows the server
to differentiate between clients and send messages
individually. Now the server just needs to know a client's
ID in order to send them a message individually.
-The join_room and create_room functionality did not differ
too much from the previous HTTP and AJAX versions. They had
a similar structure, but now the server is using websockets.
Additionally, the ClientState is now formed on connection, rather
than when a client joins/creates a room. So these join_room and
create_room functions were updated accordingly.
-The clientID and a corresponding ClientState is now generated
on connection. This removes a large headache of Component-Component
communication that would have been necessary for ClientIDs
generated on the Client side.
-New terminology: Similar Clients. These are clients within the
same room as the currentClient making WS requests. This allows
for easier function naming and commenting.