Add support for game over and killing players.
Conditional checks in onMessage added to look for
players being killed by the ghoul and to let players
know when the game is over.
Additionally, a new class GameState has been created
to keep track of ClientStates and game properties. This
object and ES6 Map objects will be used more frequently
in upcoming changes to improve efficiency of ClientState
lookups, manage each clientside game, and to clean up the
code.